requirements.txt
aiofiles==24.1.0
aiogram==3.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.6
kiwisolver==1.4.8
loguru==0.7.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.4.3
numpy==2.2.4
packaging==24.2
pandas==2.2.3
phonenumbers==9.0.2
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pyTelegramBotAPI==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.0
pytz==2025.2
requests==2.31.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.0
win32_setctime==1.2.0
yarl==1.19.0
